Rename project to "Etch.OrchardCore.Blocks"
Going to take this on at a different angle where the core fundamentals
will be to support a block based editor (with the use of a `BlockField`),
which can be extended to support different editors, starting with
editor.js.

Next step will be to move the editor.js specific code in to it's own
feature.