Add ability to link to other content items
When highlighting some content in order to make it a hyperlink, typing
the link will also search content items whose display text matches the
entered text. When the user clicks on the search results, the URL for
the content item will be set as the link in the anchor element.

The content types included in the search can be configured in the block
field settings.