Use media path to generate image URLs
When rendering an image block, the URL to the image is the URL for the
site that the block was created, which means if the content is
migrated or included in a recipe then it can cause broken images.
The image block has been refactored to make use of `mediaPath`, which
is stored on the block when selecting a media item from the media
library. This `mediaPath` is passed through to
`IMediaFileStore.MapPathToPublicUrl`. If no `mediaPath` is present then
it will default to using `url`.