Always return shape on block field/part but limit to "Detail"
Currently the block field/part won't return a shape when the display
type isn't "Detail". This restricts the ability to display the contents
of a block field/part on a different display type. Resolution is to
always return a shape but only specify a location for the "Detail"
display type. This way the developers can use the `placement.json` to
render the shape on different display types.